+++ internal/resources/container.go
@@ -59,16 +59,20 @@ func ApplyOptionsToContainers(
 
 func ContainerWithPodSecurityStandardRestricted() ContainerOption {
 	return func(c *corev1.Container) {
-		c.SecurityContext = &corev1.SecurityContext{
-			AllowPrivilegeEscalation: ptr.To(false),
-			// RunAsNonRoot:             ptr.To(true),
-			// RunAsUser:                ptr.To[int64](43690),
-			SeccompProfile: &corev1.SeccompProfile{
-				Type: corev1.SeccompProfileTypeRuntimeDefault,
-			},
-			Capabilities: &corev1.Capabilities{
-				Drop: []corev1.Capability{"ALL"},
-			},
+		if c.SecurityContext == nil {
+			c.SecurityContext = &corev1.SecurityContext{}
+		}
+		c.SecurityContext.AllowPrivilegeEscalation = ptr.To(false)
+		// Using 0 here to avoid issues with images that don't have a non-root user
+		// This is particularly important for handling git repository, since
+		// git will refuse to operate if the user of the git command does not match
+		// the owner of the .git directory/files.
+		c.SecurityContext.RunAsUser = ptr.To[int64](0)
+		c.SecurityContext.Capabilities = &corev1.Capabilities{
+			Drop: []corev1.Capability{"ALL"},
+		}
+		c.SecurityContext.SeccompProfile = &corev1.SeccompProfile{
+			Type: corev1.SeccompProfileTypeRuntimeDefault,
 		}
 	}
 }