+++ internal/webhook/v1beta1/search_webhook.go
@@ -12,6 +12,7 @@ import (
 	"context"
 	"fmt"
 
+	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -73,7 +74,20 @@ func (v *SearchCustomValidator) ValidateUpdate(ctx context.Context, oldObj, newO
 	if !ok {
 		return nil, fmt.Errorf("expected a Search object for the newObj but got %T", newObj)
 	}
+	oldSearch, ok := oldObj.(*ocularcrashoverriderunv1beta1.Search)
+	if !ok {
+		return nil, fmt.Errorf("expected a Search object for the oldObj but got %T", oldObj)
+	}
 	searchlog.Info("validating Search resource update", "name", search.GetName())
+
+	if oldSearch.Spec.ServiceAccountNameOverride != "" && oldSearch.Spec.ServiceAccountNameOverride != search.Spec.ServiceAccountNameOverride {
+		return nil, apierrors.NewInvalid(
+			schema.GroupKind{Group: "ocular.crashoverride.run", Kind: "Search"},
+			search.Name,
+			field.ErrorList{
+				field.Invalid(field.NewPath("spec").Child("serviceAccountNameOverride"), search.Spec.ServiceAccountNameOverride, "serviceAccountNameOverride cannot be changed once set"),
+			})
+	}
 
 	return nil, validateSearch(ctx, v.c, search)
 }
@@ -109,6 +123,21 @@ func validateSearch(ctx context.Context, c client.Client, search *ocularcrashove
 			return fmt.Errorf("error fetching crawler %s/%s: %w", namespace, search.Namespace, err)
 		}
 		allErrs = append(allErrs, field.NotFound(field.NewPath("spec").Child("crawlerRef").Child("name"), fmt.Sprintf("%s/%s", search.Spec.CrawlerRef.Namespace, search.Spec.CrawlerRef.Name)))
+	}
+
+	if search.Spec.ServiceAccountNameOverride != "" {
+		var serviceAccount corev1.ServiceAccount
+		err = c.Get(ctx, client.ObjectKey{
+			Name:      search.Spec.ServiceAccountNameOverride,
+			Namespace: search.Namespace,
+		}, &serviceAccount)
+
+		if err != nil {
+			if !apierrors.IsNotFound(err) {
+				return fmt.Errorf("error fetching service account %s/%s: %w", search.Namespace, search.Spec.ServiceAccountNameOverride, err)
+			}
+			allErrs = append(allErrs, field.NotFound(field.NewPath("spec").Child("serviceAccountNameOverride"), fmt.Sprintf("%s/%s", search.Namespace, search.Spec.ServiceAccountNameOverride)))
+		}
 	}
 
 	paramErrs := validateSetParameters(crawler.Name,